feat: 支持为 b 站请求配置频率限制

+++ crates/bili_sync/src/config/item.rs
@@ -35,7 +35,7 @@ pub struct ConcurrentLimit {
 
 #[derive(Serialize, Deserialize)]
 pub struct RateLimit {
-    pub limit: u64,
+    pub limit: usize,
     pub duration: u64,
 }
